♻️ refactor: BCCDデザインを取り入れた構成に修正
