+++ src/components/base/Icons/props.ts
