+++ src/components/base/Image/index.stories.tsx
