+++ src/components/base/Label/index.stories.tsx
