+++ src/components/common/Accordion/styles.css.ts
