+++ src/components/common/Dialog/styles.css.ts
